moved interfaces to type declaration file
